Move gohai logic from the metadata payload to gohai
