--- pkg/gohai/.gitignore
@@ -1,7 +1,0 @@
-.DS_Store
-*.test
-/.vagrant
-/build
-/vendor
-gohai
-*.swp
