--- pkg/gohai/Gopkg.lock
@@ -1,100 +1,0 @@
-# This file is autogenerated, do not edit; changes may be undone by the next 'dep ensure'.
-
-
-[[projects]]
-  branch = "next"
-  digest = "1:80f75788f4df66bdfb52fe5c7714878b384ec172e9382ae01c809fe25c028e9b"
-  name = "github.com/StackExchange/wmi"
-  packages = ["."]
-  pruneopts = "UT"
-  revision = "9f32b5905fd6ce7384093f9d048437e79f7b4d85"
-
-[[projects]]
-  digest = "1:50e893a85575fa48dc4982a279e50e2fd8b74e4f7c587860c1e25c77083b8125"
-  name = "github.com/cihub/seelog"
-  packages = ["."]
-  pruneopts = "UT"
-  revision = "d2c6e5aa9fbfdd1c624e140287063c7730654115"
-  version = "v2.6"
-
-[[projects]]
-  digest = "1:ffe9824d294da03b391f44e1ae8281281b4afc1bdaa9588c9097785e3af10cec"
-  name = "github.com/davecgh/go-spew"
-  packages = ["spew"]
-  pruneopts = "UT"
-  revision = "8991bc29aa16c548c550c7ff78260e27b9ab7c73"
-  version = "v1.1.1"
-
-[[projects]]
-  digest = "1:1ae488e919756729db69becc130be75a4820dfdd0775e430a85f726a84ce14ab"
-  name = "github.com/go-ole/go-ole"
-  packages = [
-    ".",
-    "oleutil",
-  ]
-  pruneopts = "UT"
-  revision = "de8695c8edbf8236f30d6e1376e20b198a028d42"
-
-[[projects]]
-  digest = "1:0028cb19b2e4c3112225cd871870f2d9cf49b9b4276531f03438a88e94be86fe"
-  name = "github.com/pmezard/go-difflib"
-  packages = ["difflib"]
-  pruneopts = "UT"
-  revision = "792786c7400a136282c1664665ae0a8db921c6c2"
-  version = "v1.0.0"
-
-[[projects]]
-  digest = "1:0dc15f8c051646249461ab843ed26b0126c528032ad8f0df2d291917ae502c12"
-  name = "github.com/shirou/gopsutil"
-  packages = [
-    "cpu",
-    "host",
-    "internal/common",
-    "mem",
-    "net",
-    "process",
-  ]
-  pruneopts = "UT"
-  revision = "e8f7a95747d711f34ddfe9dd9b825a84bd059dec"
-  version = "v2.0.0"
-
-[[projects]]
-  branch = "master"
-  digest = "1:99c6a6dab47067c9b898e8c8b13d130c6ab4ffbcc4b7cc6236c2cd0b1e344f5b"
-  name = "github.com/shirou/w32"
-  packages = ["."]
-  pruneopts = "UT"
-  revision = "bb4de0191aa41b5507caa14b0650cdbddcd9280b"
-
-[[projects]]
-  digest = "1:972c2427413d41a1e06ca4897e8528e5a1622894050e2f527b38ddf0f343f759"
-  name = "github.com/stretchr/testify"
-  packages = ["assert"]
-  pruneopts = "UT"
-  revision = "ffdc059bfe9ce6a4e144ba849dbedead332c6053"
-  version = "v1.3.0"
-
-[[projects]]
-  branch = "master"
-  digest = "1:228f27b6ae6aa190a8082e78c76870062506763cae2526543dc8f7ac9bea0947"
-  name = "golang.org/x/sys"
-  packages = [
-    "unix",
-    "windows",
-    "windows/registry",
-  ]
-  pruneopts = "UT"
-  revision = "77cc2087c03b9062d8eb49dd60e665cbd7c1cde2"
-
-[solve-meta]
-  analyzer-name = "dep"
-  analyzer-version = 1
-  input-imports = [
-    "github.com/cihub/seelog",
-    "github.com/shirou/gopsutil/mem",
-    "github.com/shirou/gopsutil/process",
-    "github.com/stretchr/testify/assert",
-    "golang.org/x/sys/windows/registry",
-  ]
-  solver-name = "gps-cdcl"
-  solver-version = 1
