+++ pkg/gohai/gohai_test.go
@@ -1,9 +1,9 @@
-// This file is licensed under the MIT License.
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright © 2015 Kentaro Kuribayashi <kentarok@gmail.com>
-// Copyright 2014-present Datadog, Inc.
+// Copyright 2016-present Datadog, Inc.
 
-package main
+package gohai
 
 import (
 	"testing"
@@ -11,10 +11,46 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestSelectedCollectors_String(t *testing.T) {
-	sc := &SelectedCollectors{
-		"foo": struct{}{},
-		"bar": struct{}{},
-	}
-	assert.Equal(t, "[bar foo]", sc.String())
+func TestGetPayload(t *testing.T) {
+	gohai := GetPayload(false)
+
+	assert.NotNil(t, gohai.Gohai.CPU)
+	assert.NotNil(t, gohai.Gohai.FileSystem)
+	assert.NotNil(t, gohai.Gohai.Memory)
+	assert.NotNil(t, gohai.Gohai.Network)
+	assert.NotNil(t, gohai.Gohai.Platform)
+}
+
+func TestGetPayloadContainerized(t *testing.T) {
+	t.Setenv("DOCKER_DD_AGENT", "true")
+
+	detectDocker0()
+	oldDocker0Detected := docker0Detected
+	docker0Detected = false
+	defer func() { docker0Detected = oldDocker0Detected }()
+
+	gohai := GetPayload(true)
+
+	assert.NotNil(t, gohai.Gohai.CPU)
+	assert.NotNil(t, gohai.Gohai.FileSystem)
+	assert.NotNil(t, gohai.Gohai.Memory)
+	assert.Nil(t, gohai.Gohai.Network)
+	assert.NotNil(t, gohai.Gohai.Platform)
+}
+
+func TestGetPayloadContainerizedWithDocker0(t *testing.T) {
+	t.Setenv("DOCKER_DD_AGENT", "true")
+
+	detectDocker0()
+	oldDocker0Detected := docker0Detected
+	docker0Detected = true
+	defer func() { docker0Detected = oldDocker0Detected }()
+
+	gohai := GetPayload(false)
+
+	assert.NotNil(t, gohai.Gohai.CPU)
+	assert.NotNil(t, gohai.Gohai.FileSystem)
+	assert.NotNil(t, gohai.Gohai.Memory)
+	assert.NotNil(t, gohai.Gohai.Network)
+	assert.NotNil(t, gohai.Gohai.Platform)
 }
