--- pkg/metadata/internal/gohai/payload.go
@@ -1,19 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-package gohai
-
-type gohai struct {
-	CPU        interface{} `json:"cpu"`
-	FileSystem interface{} `json:"filesystem"`
-	Memory     interface{} `json:"memory"`
-	Network    interface{} `json:"network"`
-	Platform   interface{} `json:"platform"`
-}
-
-// Payload handles the JSON unmarshalling of the metadata payload
-type Payload struct {
-	Gohai *gohai `json:"gohai"`
-}
